refactor: extract Page generic type with params object
