+++ src/app/catalog/secondary-housing/[subcategory]/page.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { getSecondaryHousingSubcategory } from '@/globals/api'
 import SecondaryCategory from '@/page-components/catalog/SecondaryCategory'
-import { SubcategoryPage } from '@/types/SubcategoryPage'
+import { SubcategoryPage } from '@/types/Page'
 
 async function Page({ params }: SubcategoryPage) {
   const category = await getSecondaryHousingSubcategory(params.subcategory)
