+++ src/types/Page.ts
@@ -1,0 +1,11 @@
+export type Page<TParams extends object> = {
+  params: TParams
+}
+
+export type SubcategoryPage = Page<{
+  subcategory: string
+}>
+
+export type ObjectPage = Page<{
+  object: string
+}>
