--- src/types/SubcategoryPage.ts
@@ -1,3 +1,0 @@
-export interface SubcategoryPage {
-  params: { subcategory: string }
-}
